refactor(editors/templates/enum): make editor names consistent with substation
